Fix(admin-table): new hook
